invoice-demo/requirements.txt
fastapi
uvicorn[standard]
PyJWT
python-multipart
sqlalchemy
psycopg2
